+++ src/main/java/com/lumiomedical/vault/container/definition/Definitions.java
@@ -1,10 +1,9 @@
 package com.lumiomedical.vault.container.definition;
 
 import com.lumiomedical.vault.container.Invocation;
+import com.lumiomedical.vault.container.definition.hook.PostResolvingHandler;
 
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 /**
  * @author Pierre Lecerf (pierre@noleme.com)
@@ -14,11 +13,13 @@ public class Definitions
 {
     private final Map<String, Object> variables;
     private final Map<String, ServiceDefinition> definitions;
+    private final List<PostResolvingHandler> handlers;
 
     public Definitions()
     {
         this.variables = new HashMap<>();
         this.definitions = new HashMap<>();
+        this.handlers = new ArrayList<>();
     }
 
     /**
@@ -98,6 +99,26 @@ public class Definitions
     public Definitions removeDefinition(String identifier)
     {
         this.definitions.remove(identifier);
+        return this;
+    }
+
+    /**
+     *
+     * @return
+     */
+    public List<PostResolvingHandler> getHandlers()
+    {
+        return this.handlers;
+    }
+
+    /**
+     *
+     * @param handler
+     * @return
+     */
+    public Definitions registerHandler(PostResolvingHandler handler)
+    {
+        this.handlers.add(handler);
         return this;
     }
